restructure + add homepage + add hero
