+++ src/App.jsx
@@ -1,8 +1,8 @@
 import "./App.scss";
 
 import AppRouters from "./routes/AppRoutes.jsx";
-import Header from "./components/Header.jsx";
-import Footer from './components/Footer.jsx';
+import Header from "./components/layout/Header.jsx";
+import Footer from './components/layout/Footer.jsx';
 
 
 function App() {
