+++ src/assets/scss/global.scss
@@ -52,10 +52,6 @@ body {
         & main {
             flex-grow: 1;
             background-color : var.$tiertiary;
-            display: flex;
-            align-items: center;
-            justify-content: center;
-            padding : 4rem 1.5rem;
         }
     }
 }
