--- src/layout/MainLayout.jsx
@@ -1,8 +1,0 @@
-function MainLayout({children}) {
-    return ( 
-        <>
-        </>
-    );
-}
-
-export default MainLayout;
