+++ src/pages/Home.scss
@@ -1,3 +1,127 @@
 @use "../assets/scss/variables" as var;
 @use "../assets/scss/mixins" as mxn;
 @use "../assets/scss/extends" as ext;
+
+main.main-home {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+
+
+    & section.hero {
+
+        @media screen and (min-width : var.$bp-lg) {
+            max-width: var.$bp-2xl;
+            
+            display : flex;
+            align-items: center;
+            padding : 1.5rem 1.5rem 0 1.5rem;
+        }
+
+        & div.hero-content {
+            display : flex;
+            flex-direction: column;
+            align-items: center;
+            gap : var.$gap-base;
+
+            padding : 1.5rem 2rem;
+
+            & h1 {
+                font-size: var.$txt-4xl;
+                font-weight: 800;
+                }
+
+            & h2 {
+                font-size: var.$txt-xl;
+                color : var.$gray;
+            }
+
+            & h1, h2 {
+                line-height: 1.2;
+                text-align: center;
+                max-width: 35rem;
+            }
+
+            @media screen and (max-width : var.$bp-lg) {
+                & h1, h2 {
+                    text-align: center;
+                }
+            }
+
+            @media screen and (min-width : var.$bp-lg) {
+                gap : var.$gap-xl;
+
+                & h1 {
+                    font-size: var.$txt-5xl;
+                }
+
+                & h1, h2 {
+                    text-align: start;
+                }
+            }
+
+            & div.hero-btn {
+                display : flex;
+                flex-direction: column;
+                gap: var.$gap-xl;
+                width : 100%;
+
+                & a.a-btn-secondary {
+                    @include mxn.a-btn("secondary", var.$txt-sm);
+                    width : 100%;
+                }
+    
+                & a.a-btn-primary {
+                    @include mxn.a-btn("primary", var.$txt-sm);
+                    width : 100%;
+                }
+
+                @media screen and (min-width : var.$bp-md) {
+                    flex-direction: row;
+                    justify-content: center;
+                }
+
+                @media screen and (min-width : var.$bp-lg) {
+                    justify-content: start;
+
+                    & a.a-btn-secondary,
+                    a.a-btn-primary {
+                        width: fit-content;
+                    }
+                }
+            }
+
+        }
+
+        & div.hero-img{
+            
+            & img{
+                object-fit: cover;
+                width: 100vw;
+                height: 15rem;
+                
+                @media screen and (min-width : var.$bp-md) {
+                    height : 22rem;
+                }
+
+                @media screen and (min-width : var.$bp-lg) {
+                    width : 100%;
+                    height : 35rem;
+                    -webkit-mask-image: linear-gradient(to right, transparent 0%, black 30%);
+                    mask-image: linear-gradient(to right, transparent 0%, black 30%);
+
+                    border-top-right-radius: var.$rounded-lg;;
+                    border-bottom-right-radius: var.$rounded-lg;
+                }
+            }
+        }
+
+    }
+
+    & div.content {
+        max-width: var.$bp-2xl;
+
+        display: flex;
+        flex-direction: column;
+    }
+}
