+++ src/pages/Login.scss
@@ -2,7 +2,11 @@
 @use "../assets/scss/mixins" as mxn;
 @use "../assets/scss/extends" as ext;
 
-main {
+main.main-login {
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    padding : 4rem 1.5rem;
 
     & form.login-form {
         @include mxn.shadow-thin;
