+++ src/pages/Register.jsx
@@ -30,7 +30,7 @@ function Register() {
     }, []);
 
     return ( 
-        <main>
+        <main className="main-register">
             <form onSubmit={handleRegister} className="register-form">
                 <div className="register-header">
                     <h1>Créez votre compte</h1>
